+++ test-support/define-dummy-html-input.js
@@ -1,0 +1,1 @@
+globalThis.HTMLInputElement = {prototype: {}};